requirements.txt
"Django>=3.2,<3.3"
django-dotenv==1.4.2
